Move Execute implementations to isolated files
